+++ Assets/Localization/StringTables/UI-Symbols-Localized/UI-Symbols-Localized Shared Data.asset
@@ -27,6 +27,38 @@ MonoBehaviour:
     m_Key: symbol_infinity
     m_Metadata:
       m_Items: []
+  - m_Id: 56429491015917568
+    m_Key: symbol_parenthesis_open
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500480
+    m_Key: symbol_parenthesis_close
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500481
+    m_Key: symbol_squarebrackets_open
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500482
+    m_Key: symbol_squarebrackets_close
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500483
+    m_Key: symbol_brackets_open
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500484
+    m_Key: symbol_brackets_close
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500485
+    m_Key: symbol_less
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500486
+    m_Key: symbol_greater
+    m_Metadata:
+      m_Items: []
   m_Metadata:
     m_Items: []
   m_KeyGenerator:
