+++ Assets/Localization/StringTables/UI-Symbols-Localized/UI-Symbols-Localized_en.asset
@@ -30,6 +30,38 @@ MonoBehaviour:
     m_Localized: "\u221E"
     m_Metadata:
       m_Items: []
+  - m_Id: 56429491015917568
+    m_Localized: (
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500480
+    m_Localized: )
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500481
+    m_Localized: '['
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500482
+    m_Localized: ']'
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500483
+    m_Localized: '{'
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500484
+    m_Localized: '}'
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500485
+    m_Localized: <
+    m_Metadata:
+      m_Items: []
+  - m_Id: 56429491028500486
+    m_Localized: '>'
+    m_Metadata:
+      m_Items: []
   references:
     version: 2
     RefIds: []
