--- Assets/Prefabs/UI/ButtonFloatAction.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f75f6e798531c842696ffe93364d6af8
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
