+++ Assets/Prefabs/UI/ButtonFloatClose.prefab
@@ -1,0 +1,99 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &3375701430844802949
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 2338779112939288634, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_Name
+      value: ButtonFloatClose Variant
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_Pivot.x
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_Pivot.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_AnchorMax.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_AnchorMax.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_AnchorMin.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_AnchorMin.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_SizeDelta.x
+      value: 86
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_SizeDelta.y
+      value: 86
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5075516323665096844, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 0377758677301ea2b813c0ce93ee2c6d, type: 3}
