+++ Assets/Prefabs/UI/FormationCharacterSlot.prefab
@@ -39,7 +39,7 @@ RectTransform:
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
   m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: 160, y: 86}
+  m_SizeDelta: {x: 142, y: 77}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &3223717826048600751
 CanvasRenderer:
