+++ Assets/Scripts/Manager/Character/CharacterManagerPersistance.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1c177cd62ad5fa688833dd7bbd6b9aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
